in file-suite-proc-impl, moved Delimited and OptionalDelimited to function module, and removed util module
