+++ file-suite-proc-impl/src/array_expr/function/builtin/set.rs
@@ -12,15 +12,12 @@ use ::syn::{
     punctuated::Punctuated,
 };
 
-use crate::{
-    array_expr::{
-        ArrayExpr, Node,
-        function::{Call, ToCallable},
-        storage::Storage,
-        typed_value::TypedValue,
-        value_array::ValueArray,
-    },
-    util::group_help::Delimited,
+use crate::array_expr::{
+    ArrayExpr, Node,
+    function::{Call, Delimited, ToCallable},
+    storage::Storage,
+    typed_value::TypedValue,
+    value_array::ValueArray,
 };
 
 #[derive(Debug, Clone)]
